requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
ultralytics
shapely
pandas
openpyxl
reportlab
python-barcode